app/requirements.txt
streamlit==1.55.0
pandas==2.3.3
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
matplotlib==3.10.8
scipy==1.15.3
plotly==6.6.0
altair==6.0.0